+++ src/Mission/Mission/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* General availability release for module Az.Mission targeting the stable Microsoft.Mission API version 2026-04-01
 
 ## Version 0.1.0
 * First preview release for module Az.Mission